fix(migration): bounds + fidelity + registry honesty
Closes B-MIG-1..5 from the audit:

- B-MIG-1: `read_archive_to_pending` now rejects out-of-range
  schema versions (the previous shape only checked
  `> supported`). A negative or zero `schema_version` could
  only come from a malformed manifest sentinel; surface the
  rejection alongside the future-version case so both land
  on the same `ArchiveFutureVersion` error and the import
  dialog renders the "update the app" hint either way.

- B-MIG-2: `Error::ArchiveFutureVersion::found` widens from
  `i32` to `i64` so the raw manifest value reaches the wire /
  log trace verbatim. The previous `clamp(i32::MIN..i32::MAX) as i32`
  silently capped a 64-bit value at `i32::MAX`, hiding the
  real shape from support traces; the FRB envelope's
  `Display`-driven format (`found={found}`) propagates the
  full `i64` automatically.

- B-MIG-3: Drop the eight forward-declared hw-vault / password-
  gate dependency entries in `migration::registry::build_app_registry`.
  None of them register an `Artefact`; their version envelopes
  (`HW_VAULT_*` / DISK_BLOB_VERSION) run their own corruption-
  detection cascade outside the generic migration framework.
  The forward declarations were dead weight that `topo_sort`
  silently skipped. New invariant test
  `every_dependency_endpoint_is_a_registered_artefact` pins
  the rule so a future commit that re-introduces a dead edge
  fails at test time. `topo_sort` now `app_log_warn!`s on
  unrecognised endpoints instead of silent skip — the rule
  is enforced at registry build, but a runtime drift still
  leaves a support-trace breadcrumb.

- B-MIG-4: `bootstrap_schema` now stamps `PRAGMA user_version`
  only when the current value is `0` (fresh DB). The
  unconditional stamp would silently downgrade a future
  schema bump: a v2 DB opened by a v1 build would have its
  `user_version` rewritten to 1, and the next v2 build would
  mistake the file for a v1 install needing a forward
  migration that already ran. Tables stay
  `CREATE IF NOT EXISTS` so the call remains safe to re-run
  on every open.

- B-MIG-5: `parse_pending_import` now `app_log_warn!`s on
  unknown ZIP entries instead of silently dropping them. A
  forward-compat archive (a future build shipping
  `recordings_index.json` or similar) leaves a greppable
  trace; a malformed v1 archive with stray content also
  leaves a trail rather than disappearing into the void.
  The schema-version check at the caller still catches the
  "future build" case as an explicit error.

+++ rust/crates/lfs_core/src/error.rs
@@ -118,12 +118,16 @@ pub enum Error {
     Cancelled,
 
     /// `.lfs` archive's manifest reports a `schema_version` newer
-    /// than `migration::SchemaVersions::ARCHIVE`. The Display
-    /// format is parsed Dart-side into `UnsupportedLfsVersionException`,
-    /// which the import dialog renders with the "update the app"
-    /// message — keep the `found=` / `supported=` shape stable.
+    /// than `migration::SchemaVersions::ARCHIVE`, or an
+    /// out-of-range value (`<= 0`). The Display format is parsed
+    /// Dart-side into `UnsupportedLfsVersionException`, which the
+    /// import dialog renders with the "update the app" message —
+    /// keep the `found=` / `supported=` shape stable. `found` is
+    /// `i64` so the raw manifest value reaches the user / log
+    /// trace verbatim instead of being clamped to `i32::MAX` (the
+    /// audit B-MIG-2 fidelity gap).
     #[error("unsupported_archive_version: found={found}, supported={supported}")]
-    ArchiveFutureVersion { found: i32, supported: i32 },
+    ArchiveFutureVersion { found: i64, supported: i32 },
 }
 
 impl From<russh::Error> for Error {
